requirements.txt
absl-py==2.3.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
easydict==1.13
filelock==3.18.0
filterpy==1.4.5
fonttools==4.59.0
fsspec==2025.7.0
gdown==5.2.0
grpcio==1.74.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prettytable==3.16.0
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.172
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
yacs==0.1.8
